% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{alphonse}
\alias{alphonse}
\title{Generate color palettes based on popular anime.}
\usage{
alphonse(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of color palette. Options are: \code{mha}, \code{mhadeku}, \code{mhabakugo},
\code{mhashoto}, \code{jjk}, \code{demon}, \code{aot}, \code{mob}, \code{fma}, \code{naruto},
\code{stone}, \code{ff}, \code{opm}, \code{hxh}, \code{haikyuu}, \code{clover}, \code{jojo}, 
\code{madoka}, \code{pokemon}, \code{onepiece}, and \code{dbz}.}

\item{n}{Number of colors needed.}

\item{type}{Continuous or discrete.}
}
\description{
h/t to @jrcalabrese and their joycon repository, and @ciannabp and their 
inauguration repository, for inspiration and base code for making a color palette package.
}
