% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{alphonse_palettes}
\alias{alphonse_palettes}
\title{Color palettes based on popular anime.}
\format{
An object of class \code{list} of length 21.
}
\usage{
alphonse_palettes
}
\description{
Color options include: \code{mha}, \code{mhadeku}, \code{mhabakugo},
\code{mhashoto}, \code{jjk}, \code{demon},
\code{aot}, \code{mob}, \code{fma},
\code{naruto}, \code{stone}, \code{ff},
\code{opm}, \code{hxh}, \code{haikyuu}, \code{clover},
\code{jojo}, \code{madoka}, \code{pokemon}, \code{onepiece},
and \code{dbz}.
}
\keyword{datasets}
